the matrix multiplication is ok

+++ main.hpp
@@ -1,4 +1,5 @@
 #pragma once
 // using valT = double;
-using valT = long double;
+// using valT = long double;
+using valT = float;
 using funcT = valT (*)(valT);
